server/fastapi_analytics/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
firebase-admin==6.4.0
